Move to shared TA utils
