+++ libs/application/template-api-modules/src/lib/modules/templates/transport-authority/utils/index.ts
@@ -1,0 +1,1 @@
+export * from './mapVehicle'
